refactor: standardize `ControllerLevel` usage throughout the codebase
- Replaced integer-based controller levels with strongly-typed `ControllerLevel` enum.
- Updated models, repositories, and services to adopt `ControllerLevel` for improved type safety.
- Adjusted related test cases and factory logic to reflect changes.

+++ src/ScreepsDotNet.Storage.MongoRedis/Seeding/SeedDataService.cs
@@ -321,7 +321,7 @@ public sealed class SeedDataService : ISeedDataService
             Owner = SeedDataDefaults.User.Username,
             Controller = new RoomControllerDocument
             {
-                Level = (int)DefaultControllerLevel
+                Level = DefaultControllerLevel
             },
             EnergyAvailable = 500
         };
